Add the hashchange event to resizers.
With Single Page web architecture and internal routers, it's a likely scenario that many will offer up a hash change instead of a redirect. If this is not listened for it will cause for a potential miss in resizing, and after a single click the next loaded "screen" will not have the appropriately placed links.